+++ spec/support/user.rb
@@ -33,6 +33,7 @@ class UserWithAttributes < ActiveRecord::Base
   store_accessor :jparams, active: :boolean, birthday: :date, prefix: "json", suffix: "value"
   store_attribute :jparams, :inner_json, :json
   store_attribute :jdata, :salary, :integer
+  store_attribute :jdata, :hours, ActiveRecord::Type.lookup(:integer)
   store :custom, accessors: [:custom_date, price: :money_type]
 end